процесс рефакторинга начат, но я почему-то начал с конца
