+++ src/main/java/io/github/_7isenko/aegis/TODO.yml
@@ -1,0 +1,7 @@
+TODO:
+  - инкапсулировать команды
+  - сделать красивый враппер конфига
+  - брать каналы по названиям, а не по айди
+  - докуметация
+  - использовать паттерны
+  -
