+++ src/main/java/io/github/_7isenko/aegis/minecraft/UUIDHelper.java
@@ -1,4 +1,4 @@
-package io.github._7isenko.aegis;
+package io.github._7isenko.aegis.minecraft;
 
 import java.util.UUID;
 import java.io.IOException;
@@ -9,13 +9,9 @@ import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 import org.json.simple.parser.ParseException;
 
-public class UUIDGetter {
-    private static UUIDGetter instance;
+public class UUIDHelper {
 
-    private UUIDGetter() {
-    }
-
-    public UUID getUuid(String nickname) throws NullPointerException {
+    public static UUID getUuid(String nickname) throws NullPointerException {
         String url = "https://api.mojang.com/users/profiles/minecraft/" + nickname;
         try {
             @SuppressWarnings("deprecation")
@@ -29,18 +25,8 @@ public class UUIDGetter {
         throw new NullPointerException();
     }
 
-    private String addDashes(String uuid) {
+    private static String addDashes(String uuid) {
         return uuid.replaceFirst("(\\p{XDigit}{8})(\\p{XDigit}{4})(\\p{XDigit}{4})(\\p{XDigit}{4})(\\p{XDigit}+)", "$1-$2-$3-$4-$5");
     }
 
-    public static UUIDGetter getInstance() {
-        if (instance == null) {
-            instance = new UUIDGetter();
-        }
-        return instance;
-    }
-
-    public static boolean hasInstance() {
-        return instance != null;
-    }
 }
