+++ src/main/java/io/github/_7isenko/aegis/misc/StringsTool.java
@@ -1,4 +1,4 @@
-package io.github._7isenko.aegis;
+package io.github._7isenko.aegis.misc;
 
 import net.dv8tion.jda.api.entities.Member;
 
@@ -15,7 +15,7 @@ public class StringsTool {
             sb.append(i).append(". ");
             if (member != null)
                 sb.append(member.getUser().getAsTag()).append("\n");
-            else sb.append("ливнул").append("\n");
+            else sb.append("dummy#0000").append("\n");
             if (i % 20 == 0) {
                 strings.add(sb.toString());
                 sb = new StringBuilder();
