+++ src/main/resources/config.yml
@@ -1,13 +1,11 @@
 token: put_here_your_token
-server_id: 1234567890
-control_channel_id: 1234567890
-whitelist_channel_id: 1234567890
-announce_channel_id: 1234567890
-greeting_channel_id: 1234567890
-member_role_id: 1234567890
-chosen_role_id: 1234567890
+serverId: 1234567890
 
 allow_kick: true
 
-# Поставить true, если конфиг заполнен.
-filled: false
+startMessages:
+  - "*ввввд-жуужу-жббжжужу*\nРобот запущен. Рандом начинается!"
+  - "Чтооо? Опять? Ну ладно, начинаем..."
+  - "Сегодня так много новых лиц!\nРад всех видеть, надеюсь, будет много новичков!"
+  - "- Я всего лишь хочу тебе объяснить, что все люди созданы разными, с разными возможностями и разными потребностями. Глупо было бы требовать от всех людей одного и того же, это всё равно что заставлять пингвина летать или рыбу петь...\n- Петрович, заебал, начинай уже рандомить!!!"
+  - "привет фывфыв я в дс но не на серве добавишь в вайт лист?? я ждал 68 часов и не успел го перенабор пожалуйста!!! почему я в дс но не могу писать в вайтлист что мне делать??? эм я зашел на в дс , но у меня нет роли закинь в гч съемки пожалуйста!! привет фывфыв добавь в друзья вк прошууу . закрепи этот комментарий если любишь подписчиков. эмммм привет фывфыв я упал в самом начале го второй шанс пж..."
